remove forest-fil-types crate and relocate items
