+++ blockchain/consensus/fil_cns/src/validation.rs
@@ -9,10 +9,9 @@ use fil_actors_runtime_v9::runtime::DomainSeparationTag;
 use forest_beacon::{Beacon, BeaconEntry, BeaconSchedule, IGNORE_DRAND_VAR};
 use forest_blocks::{Block, BlockHeader, Tipset};
 use forest_chain_sync::collect_errs;
-use forest_fil_types::verifier::verify_winning_post;
 use forest_networks::{ChainConfig, Height};
 use forest_shim::{address::Address, randomness::Randomness, version::NetworkVersion};
-use forest_state_manager::StateManager;
+use forest_state_manager::{verify_winning_post, StateManager};
 use futures::stream::FuturesUnordered;
 use fvm_ipld_blockstore::Blockstore;
 use fvm_ipld_encoding::Cbor;
