--- types/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "forest_fil_types"
-description = "Filecoin types used in Forest."
-license.workspace = true
-repository.workspace = true
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-filecoin-proofs-api = { workspace = true, default-features = false }
-forest_shim.workspace = true
-fvm_ipld_encoding.workspace = true
-fvm_shared = { workspace = true, default-features = false }
-
-[dev-dependencies]
