--- types/src/lib.rs
@@ -1,4 +1,0 @@
-// Copyright 2019-2023 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-pub mod verifier;
