+++ utils/forest_utils/Cargo.toml
@@ -15,10 +15,13 @@ blake2b_simd.workspace = true
 cid.workspace = true
 const_format = "0.2.26"
 digest.workspace = true
+filecoin-proofs-api = { workspace = true, default-features = false }
+forest_shim.workspace = true
 futures.workspace = true
 fvm_ipld_blockstore.workspace = true
 fvm_ipld_encoding.workspace = true
 fvm_ipld_encoding3.workspace = true
+fvm_shared = { workspace = true, default-features = false }
 hyper-rustls.workspace = true
 hyper.workspace = true
 libc = "0.2"
